# API_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: Seven Steps to Poem API
  description: AI Agent system for automated business problem solving using McKinsey's 7-step methodology
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@sevenstepstopoem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sevenstepstopoem.com/v1
    description: Production server
  - url: https://staging-api.sevenstepstopoem.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - BearerAuth: []
paths:
  /problems:
    post:
      summary: Create a new problem analysis request
      description: Submit a business problem for AI agent analysis
      tags:
        - Problems
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - raw_text
                - submitter
              properties:
                raw_text:
                  type: string
                  description: Natural language description of the business problem
                  example: "Our SaaS customer churn rate has increased 20% over the past 3 months. We need to reduce it by 5% within 90 days."
                submitter:
                  type: string
                  format: email
                  description: Email address of the person submitting the problem
                  example: "john.doe@company.com"
                attachments:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Supporting files (PDF, CSV, XLSX)
                urgency:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
                  description: Priority level of the problem
                metadata:
                  type: object
                  description: Additional context and metadata
                  properties:
                    industry:
                      type: string
                      example: "SaaS"
                    company_size:
                      type: string
                      enum: [startup, small, medium, large, enterprise]
                    budget_range:
                      type: string
                      example: "$10K-50K"
      responses:
        '201':
          description: Problem created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '429':
          description: Rate limit exceeded

    get:
      summary: List all problems for the authenticated user
      description: Retrieve a paginated list of problems
      tags:
        - Problems
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [created, framing, tree_generation, prioritizing, planning, analyzing, synthesizing, presenting, completed, failed]
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, updated_at, urgency]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: List of problems
          content:
            application/json:
              schema:
                type: object
                properties:
                  problems:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProblemSummary'
                  pagination:
                    $ref: '#/components/schemas/PaginationInfo'
...
  /problems/{problem_id}/frame:
    get:
      summary: Get problem framing results
      description: Retrieve the structured problem frame
      tags:
        - Analysis Steps
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Problem frame data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemFrame'
        '404':
          description: Problem not found
        '409':
          description: Framing not yet completed

    post:
      summary: Update problem frame
      description: Provide clarifications or updates to the problem frame
      tags:
        - Analysis Steps
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProblemFrameUpdate'
      responses:
        '200':
          description: Problem frame updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemFrame'
...
  /problems/{problem_id}/executions/{execution_id}:
    get:
      summary: Get execution status
      description: Check the status of a running analysis step
      tags:
        - Execution
      parameters:
        - name: problem_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Execution status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStatus'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ProblemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [created, framing_pending]
        created_at:
          type: string
          format: date-time
        estimated_completion:
          type: string
          format: date-time
    ProblemSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        urgency:
          type: string
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
...
    ProblemFrame:
      type: object
      properties:
        goal:
          type: string
          example: "Reduce SaaS customer churn rate by 5% within 90 days"
        scope:
          type: object
          properties:
            include:
              type: array
              items:
                type: string
              example: ["North American paid customers", "Enterprise tier"]
            exclude:
              type: array
              items:
                type: string
              example: ["Free trial users", "Beta customers"]
        kpis:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              baseline:
                type: number
              target:
                type: number
              window:
                type: string
              unit:
                type: string
        stakeholders:
          type: array
          items:
            type: string
        assumptions:
          type: array
          items:
            type: string
        constraints:
          type: array
          items:
            type: string
        confidence:
          type: string
          enum: [low, medium, high]
        clarifying_questions:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              category:
                type: string
              required:
                type: boolean
    ProblemFrameUpdate:
      type: object
      properties:
        clarifications:
          type: object
          additionalProperties:
            type: string
        updated_assumptions:
          type: array
          items:
            type: string
        additional_constraints:
          type: array
          items:
            type: string
...
    ExecutionStatus:
      type: object
      properties:
        execution_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        progress_percentage:
          type: integer
          minimum: 0
          maximum: 100
        current_activity:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        estimated_remaining:
          type: integer
          description: Estimated remaining time in seconds
        logs:
          type: array
          items:
            type: object
            properties:
              timestamp:
                type: string
                format: date-time
              level:
                type: string
                enum: [debug, info, warning, error]
              message:
                type: string
        error_details:
          type: object
          properties:
            error_code:
              type: string
            error_message:
              type: string
            stack_trace:
              type: string
...
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
        error_message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
        has_next:
          type: boolean
        has_previous:
          type: boolean
...